requirements.txt
streamlit
Pillow
pandas
numpy
matplotlib
seaborn
joblib
lightgbm
category-encoders
scikit-learn==1.6.1